% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{binomial_price}
\alias{binomial_price}
\title{Binomial tree options pricing}
\usage{
binomial_price(S, K, T, r, q, sigma, N, is_call, is_american)
}
\arguments{
\item{S}{numeric scalar: current price of the underlying asset}

\item{K}{numeric scalar: strike price of the option}

\item{T}{numeric scalar: time to expiration of the option (in years)}

\item{r}{numeric scalar: risk-free interest rate (annualized)}

\item{q}{numeric scalar: continuous dividend yield of the underlying asset (annualized)}

\item{sigma}{numeric scalar: volatility of the underlying asset (annualized)}

\item{N}{integer scalar: number of steps in the binomial tree}

\item{is_call}{integer scalar: if 1, price a call option; if 0, price a put option}

\item{is_american}{integer scalar: if 1, price an American option; if 0, price a European option}
}
\value{
Returns a numeric scalar in the range (0, +inf) representing the price of the option.
}
\description{
Computes the theoretical price of an option using a binomial tree model.
}
\examples{
binomial_price(100, 100, 1, 0.05, 0.02, 0.2, 100, 1, 0)
stopifnot(binomial_price(100, 100, 1, 0.05, 0.02, 0.2, 100, 0, 1) > 0)
}
