% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{bs_price}
\alias{bs_price}
\title{Black-Scholes option pricing}
\usage{
bs_price(S, K, T, r, sigma, is_call)
}
\arguments{
\item{S}{numeric scalar: current price of the underlying asset}

\item{K}{numeric scalar: strike price of the option}

\item{T}{numeric scalar: time to expiration of the option (in years)}

\item{r}{numeric scalar: risk-free interest rate (annualized)}

\item{sigma}{numeric scalar: volatility of the underlying asset (annualized)}

\item{is_call}{integer scalar: if 1, price a call option; if 0, price a put option}
}
\value{
Returns a numeric scalar in the range (0, +inf) representing the price of the option.
}
\description{
Computes the theoretical price of an option using the Black-Scholes formula.
}
\examples{
bs_price(100, 100, 1, 0.05, 0.2, 1)
stopifnot(bs_price(100, 100, 1, 0.05, 0.2, 0) > 0)
}
