% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical.R
\name{integrate_gauss}
\alias{integrate_gauss}
\title{Gauss quadrature integratinon}
\usage{
integrate_gauss(f, a, b, n_points = 10L)
}
\arguments{
\item{f}{function: scalar -> scalar whose integral we seek}

\item{a}{numeric: bracket left endpoint}

\item{b}{numeric: bracket right endpoint}

\item{n_points}{integer: number of quadrature points (default 10)}
}
\value{
Returns a numeric scalar representing the approximate integral of f over [a, b].
}
\description{
Integrates f over [a, b] using Gauss quadrature.
}
\examples{
integrate_gauss(function(x) x^2, 0, 1, n_points = 5L)
}
