% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{mc_price_european}
\alias{mc_price_european}
\title{European options pricing via Monte Carlo simulation}
\usage{
mc_price_european(S, K, T, r, sigma, N_paths, is_call)
}
\arguments{
\item{S}{numeric scalar: current price of the underlying asset}

\item{K}{numeric scalar: strike price of the option}

\item{T}{numeric scalar: time to expiration of the option (in years)}

\item{r}{numeric scalar: risk-free interest rate (annualized)}

\item{sigma}{numeric scalar: volatility of the underlying asset (annualized)}

\item{N_paths}{integer scalar: number of simulated paths}

\item{is_call}{integer scalar: if 1, price a call option; if 0, price a put option}
}
\value{
Returns a tuple of numeric scalars, both in the range (0, +inf), representing the price of the option and the standard error of the price estimate.
}
\description{
Computes the theoretical price of a European option using Monte Carlo with Geometric Brownian Motion paths.
}
\examples{
mc_price_european(100, 100, 1, 0.05, 0.2, 10000, 1)
stopifnot(mc_price_european(100, 100, 1, 0.05, 0.2, 1000, 1)[1] > 0)
}
