% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatility.R
\name{vol_parkinson}
\alias{vol_parkinson}
\title{Parkinson volatility calculation}
\usage{
vol_parkinson(high, low, n)
}
\arguments{
\item{high}{numeric vector: daily high prices}

\item{low}{numeric vector: daily low prices}

\item{n}{integer scalar: number of observations}
}
\value{
Returns a numeric scalar representing annualized volatility.
}
\description{
Computes annualized volatility from high/low prices.
}
\examples{
vol_parkinson(c(102,101,104), c(99,98,100), 3)
}
