% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical.R, R/pricing.R
\docType{package}
\name{Volaris-package}
\alias{Volaris-package}
\title{Volaris: High-performance library for quantitative finance}
\description{
Provides time-efficient pricing of derivatives, volatility estimation, a stochastic simulation engine and calculation of risk metrics via a compiled C/C++ core.

Provides time-efficient pricing of derivatives, volatility estimation, a stochastic simulation engine and calculation of risk metrics via a compiled C/C++ core.
}
\author{
\strong{Maintainer}: Barbara Gawlik \email{basia.m.gawlik@gmail.com}

}
